fix minting issue
