+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "@story-protocol/core-sdk": "^0.0.1-beta-rc.3",
+    "ethers": "5.4.0",
     "ts-node": "^10.9.2",
     "viem": "^1.21.4"
   },
